Part of a patch from Albert Lee <trisk at forkgnu.org>:
- Clean up configure tests and check if isfinite is allowed by the C++ compiler
- Use <clocale> instead of <locale> for LC_ALL and std::setlocale
- Remove duplicates from GNASHRC list
- Remove most OS-specific content in prcpucfg-glibc.h
- Improve architecture and compiler support in prcpucfg-glibc.h
- Use std:: prefixes consistently
- Use putenv (SUSv2) instead of setenv to modify GNASHRC
- Prepend rather than append new GNASHRC entries
- Use gnash::StringNoCaseEqual instead of strcasecmp
- Fix quotes in agg autoconf macro
- Prevent pthreads check from mangling CXXFLAGS

+++ backend/Renderer_ogl.cpp
@@ -789,8 +789,8 @@ public:
 
     gluOrtho2D(x0, x1, y0, y1);
     
-    _width  = fabsf(x1 - x0);
-    _height = fabsf(y1 - y0);
+    _width  = std::fabs(x1 - x0);
+    _height = std::fabs(y1 - y0);
 
     glScalef(static_cast<float>(twipsToPixels(_width)) /
     static_cast<float>(viewport_width),
@@ -1309,9 +1309,9 @@ public:
          LOG_ONCE( log_unimpl(_("Unidirectionally scaled strokes in OGL renderer")) );
       }
       
-      float stroke_scale = fabsf(mat.get_x_scale()) + fabsf(mat.get_y_scale());
+      float stroke_scale = std::fabs(mat.get_x_scale()) + std::fabs(mat.get_y_scale());
       stroke_scale /= 2.0f;
-      stroke_scale *= (fabsf(_xscale) + fabsf(_yscale)) / 2.0f;
+      stroke_scale *= (std::fabs(_xscale) + std::fabs(_yscale)) / 2.0f;
       width *= stroke_scale;
       width = twipsToPixels(width);
